relocate/rename mojo
