+++ tycho-its/projects/p2mavensite/consumer/META-INF/MANIFEST.MF
