+++ tycho-its/projects/p2mavensite/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project>
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.eclipse.tycho.extras.it</groupId>
+	<groupId>org.eclipse.tycho.it</groupId>
 	<artifactId>mvn-p2-site-parent</artifactId>
 	<version>1.0.0</version>
 	<packaging>pom</packaging>
