+++ tycho-its/projects/p2mavensite/producer/pom.xml
@@ -2,7 +2,7 @@
 <project>
 
 	<parent>
-		<groupId>org.eclipse.tycho.extras.it</groupId>
+		<groupId>org.eclipse.tycho.it</groupId>
 		<artifactId>mvn-p2-site-parent</artifactId>
 		<version>1.0.0</version>
 	</parent>
@@ -12,8 +12,8 @@
 	<build>
 		<plugins>
 			<plugin>
-				<groupId>org.eclipse.tycho.extras</groupId>
-				<artifactId>tycho-p2-extras-plugin</artifactId>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-p2-repository-plugin</artifactId>
 				<version>${tycho-version}</version>
 				<executions>
 					<execution>
@@ -25,7 +25,7 @@
 						<id>maven-p2-site</id>
 						<phase>package</phase>
 						<goals>
-							<goal>maven-p2-site</goal>
+							<goal>assemble-p2-maven-site</goal>
 						</goals>
 					</execution>
 				</executions>
