+++ tycho-p2/tycho-p2-repository-plugin/pom.xml
@@ -40,6 +40,11 @@
 			<artifactId>tycho-p2-facade</artifactId>
 			<version>${project.version}</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.tycho</groupId>
+			<artifactId>sisu-equinox-launching</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 	</dependencies>
 
 </project>
